feat(excel)!: v1.0 API breaking changes, test-infra overhaul, README expansion
BREAKING CHANGES (approved for the v1.0 push):
- Application.Visible/DisplayAlerts/ScreenUpdating return (bool, error)
- Charts.Add takes functional options (ChartAt/ChartSize; xlwings defaults)
- Vector1D/Vector2D aliases removed (use Vector/Grid)
- Range.UnMerge renamed Unmerge (COM member name preserved internally)
- Application.Hwnd returns uintptr; PID returns uint32 (DWORD)
- Range.AutoFit fits both column width and row height (xlwings parity)
- Workbook.SaveAs/Close gain functional options (SaveFileFormat,
  SavePassword, CloseSaveChanges) with a FileFormat const set

Test infrastructure:
- COM tests gated behind windows && excel_integration; go test ./...
  no longer boots Excel (core arena/expression tests converted to
  Scripting.Dictionary/FSO; TestRGB split out untagged)
- internal/testutil: two-tier process cleanup (graceful Quit 5s, then
  TerminateProcess, then confirm) + PIDFromHwnd
- excel/harness_test.go: shared withApp/withBook/withSheet harness
  replacing three divergent copies; Taskfile integration target runs ./...

Docs: README +~170 lines of runnable examples (Books.Open, Range
essentials, Charts, Pictures/Shapes, Names, OptionedRange.Set,
Missing()/Error()), updated to the new signatures; AGENTS.md synced.

Verified: go test ./... Excel-free, full and targeted
-tags=excel_integration runs pass with zero leaked EXCEL.EXE.

+++ excel/win32.go
@@ -20,15 +20,16 @@ var (
 )
 
 // pidFromHwnd returns the PID owning the given top-level window. Returns an
-// error if the Win32 call fails or the window has gone away.
-func pidFromHwnd(hwnd int32) (int32, error) {
+// error if the Win32 call fails or the window has gone away. The PID is an
+// unsigned DWORD (uint32).
+func pidFromHwnd(hwnd uintptr) (uint32, error) {
 	var pid uint32
 	ret, _, callErr := procGetWindowThreadProcessId.Call(
-		uintptr(hwnd),
+		hwnd,
 		uintptr(unsafe.Pointer(&pid)),
 	)
 	if ret == 0 {
 		return 0, fmt.Errorf("GetWindowThreadProcessId failed for hwnd 0x%x: %v", hwnd, callErr)
 	}
-	return int32(pid), nil
+	return pid, nil
 }
